# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SolarIncentives API",
        "description": "Clean-energy incentive platform — projects, multi-step applications, status workflows, documents, notifications.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/applications/{application}/documents": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document to an application",
                "operationId": "9aacefae69249bf3fdffb1e08c0c0255",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "type"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "pdf/jpg/jpeg/png, ≤10MB",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uploaded (with signed download_url)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/documents": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document to a project",
                "operationId": "1f0e1a8f8b54ed695107a0d91092ea41",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "type"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "pdf/jpg/jpeg/png, ≤10MB",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uploaded"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/documents/{document}/download": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Stream a private document (temporary signed URL — no token)",
                "operationId": "84f9069fe7a3e13c0f6b187b3f6f72ac",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File stream"
                    },
                    "403": {
                        "description": "Invalid/expired signature"
                    }
                }
            }
        },
        "/documents/{document}": {
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document (row + stored file)",
                "operationId": "0e30b6b5f7f3d6eb00a98924ec392c6c",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/projects/{project}/transition": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Transition project status (state machine)",
                "operationId": "2baad50257a0ca10f891ef3bbc46a168",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "to"
                                ],
                                "properties": {
                                    "to": {
                                        "type": "string",
                                        "enum": [
                                            "submitted",
                                            "in_review",
                                            "approved",
                                            "installed",
                                            "closed",
                                            "rejected"
                                        ]
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated project"
                    },
                    "422": {
                        "description": "Illegal transition"
                    },
                    "403": {
                        "description": "Wrong role"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}